requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
attrs=23.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
blinker=1.8.2=pypi_0
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.6.2=py39h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.2.1=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.12.1=pypi_0
dacite=1.8.1=pypi_0
dash=2.17.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.8.2=pypi_0
decorator=5.1.1=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
faiss-gpu=1.7.2=pypi_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
flask=3.0.3=pypi_0
flatbuffers=24.3.25=pypi_0
flow-vis=0.1=pypi_0
fonttools=4.53.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.6.0=pypi_0
gast=0.6.0=pypi_0
gdown=5.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.30.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.64.1=pypi_0
gsplat=1.0.0=pypi_0
h5py=3.10.0=pypi_0
hloc=1.5=dev_0
icecream=2.1.3=pypi_0
idna=3.7=py39h06a4308_0
imageio=2.34.2=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipdb=0.13.13=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.3=pypi_0
itsdangerous=2.2.0=pypi_0
jaxtyping=0.2.31=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=py39h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
keopscore=2.2.3=pypi_0
keras=3.4.1=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.3=pypi_0
kornia-rs=0.1.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=18.1.1=pypi_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.10.0.4=0
libcurand=10.3.6.39=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lightglue=0.0=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
ml-dtypes=0.4.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
namex=0.0.8=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
ninja=1.11.1.1=pypi_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
oauthlib=3.2.2=pypi_0
open3d=0.18.0+fcc396e=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.14=h5eee18b_0
opt-einsum=3.3.0=pypi_0
optree=0.12.1=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.3.0=py39h5eee18b_0
pip=24.0=py39h06a4308_0
platformdirs=4.2.2=pypi_0
plotly=5.22.0=pypi_0
plydata=0.4.3=pypi_0
plyfile=1.0.3=pypi_0
point-cloud-utils=0.30.4=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=4.25.3=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pybind11=2.13.6=pypi_0
pycolmap=0.6.1=pypi_0
pyglet=1.5.29=pypi_0
pygments=2.18.0=pypi_0
pyhocon=0.3.59=pypi_0
pykeops=2.2.3=pypi_0
pyparsing=2.4.7=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytorch-ranger=0.1.1=pypi_0
pytz=2024.1=pypi_0
pywavelets=1.6.0=pypi_0
pyyaml=6.0.1=py39h5eee18b_0
pyzmq=26.0.3=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.2=py39h06a4308_0
requests-oauthlib=2.0.0=pypi_0
retrying=1.3.4=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=69.5.1=py39h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.4.2=pypi_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.17.0=pypi_0
tensorflow-io-gcs-filesystem=0.37.1=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.6.18=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=pypi_0
torch=1.13.1+cu116=pypi_0
torch-optimizer=0.3.0=pypi_0
torchaudio=0.13.1+cu116=pypi_0
torchtriton=2.1.0=py39
torchvision=0.14.1+cu116=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
trimesh=4.2.2=pypi_0
typeguard=2.13.3=pypi_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=py39h06a4308_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39h06a4308_0
widgetsnbextension=4.0.11=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
